Add long validation contract
Adding long validation contract.

+++ Flunt/Validations/LongValidationContract.cs
@@ -3,7 +3,7 @@
     public partial class Contract
     {
         #region IsGreaterThan
-        public Contract IsGreaterThan(decimal val, int comparer, string property, string message)
+        public Contract IsGreaterThan(decimal val, long comparer, string property, string message)
         {
             if ((double)val <= comparer)
                 AddNotification(property, message);
@@ -11,7 +11,7 @@
             return this;
         }
 
-        public Contract IsGreaterThan(double val, int comparer, string property, string message)
+        public Contract IsGreaterThan(double val, long comparer, string property, string message)
         {
             if (val <= comparer)
                 AddNotification(property, message);
@@ -19,7 +19,7 @@
             return this;
         }
 
-        public Contract IsGreaterThan(float val, int comparer, string property, string message)
+        public Contract IsGreaterThan(float val, long comparer, string property, string message)
         {
             if (val <= comparer)
                 AddNotification(property, message);
@@ -27,7 +27,7 @@
             return this;
         }
 
-        public Contract IsGreaterThan(long val, int comparer, string property, string message)
+        public Contract IsGreaterThan(int val, long comparer, string property, string message)
         {
             if (val <= comparer)
                 AddNotification(property, message);
@@ -35,7 +35,7 @@
             return this;
         }
 
-        public Contract IsGreaterThan(int val, int comparer, string property, string message)
+        public Contract IsGreaterThan(long val, long comparer, string property, string message)
         {
             if (val <= comparer)
                 AddNotification(property, message);
@@ -45,7 +45,7 @@
         #endregion
 
         #region IsGreaterOrEqualsThan
-        public Contract IsGreaterOrEqualsThan(decimal val, int comparer, string property, string message)
+        public Contract IsGreaterOrEqualsThan(decimal val, long comparer, string property, string message)
         {
             if ((double)val < comparer)
                 AddNotification(property, message);
@@ -53,7 +53,7 @@
             return this;
         }
 
-        public Contract IsGreaterOrEqualsThan(double val, int comparer, string property, string message)
+        public Contract IsGreaterOrEqualsThan(double val, long comparer, string property, string message)
         {
             if (val < comparer)
                 AddNotification(property, message);
@@ -61,7 +61,7 @@
             return this;
         }
 
-        public Contract IsGreaterOrEqualsThan(float val, int comparer, string property, string message)
+        public Contract IsGreaterOrEqualsThan(float val, long comparer, string property, string message)
         {
             if (val < comparer)
                 AddNotification(property, message);
@@ -69,7 +69,7 @@
             return this;
         }
 
-        public Contract IsGreaterOrEqualsThan(long val, int comparer, string property, string message)
+        public Contract IsGreaterOrEqualsThan(int val, long comparer, string property, string message)
         {
             if (val < comparer)
                 AddNotification(property, message);
@@ -77,7 +77,7 @@
             return this;
         }
 
-        public Contract IsGreaterOrEqualsThan(int val, int comparer, string property, string message)
+        public Contract IsGreaterOrEqualsThan(long val, long comparer, string property, string message)
         {
             if (val < comparer)
                 AddNotification(property, message);
@@ -87,7 +87,7 @@
         #endregion
 
         #region IsLowerThan
-        public Contract IsLowerThan(decimal val, int comparer, string property, string message)
+        public Contract IsLowerThan(decimal val, long comparer, string property, string message)
         {
             if ((double)val >= comparer)
                 AddNotification(property, message);
@@ -95,7 +95,7 @@
             return this;
         }
 
-        public Contract IsLowerThan(double val, int comparer, string property, string message)
+        public Contract IsLowerThan(double val, long comparer, string property, string message)
         {
             if (val >= comparer)
                 AddNotification(property, message);
@@ -103,7 +103,7 @@
             return this;
         }
 
-        public Contract IsLowerThan(float val, int comparer, string property, string message)
+        public Contract IsLowerThan(float val, long comparer, string property, string message)
         {
             if (val >= comparer)
                 AddNotification(property, message);
@@ -111,7 +111,7 @@
             return this;
         }
 
-        public Contract IsLowerThan(long val, int comparer, string property, string message)
+        public Contract IsLowerThan(int val, long comparer, string property, string message)
         {
             if (val >= comparer)
                 AddNotification(property, message);
@@ -119,7 +119,7 @@
             return this;
         }
 
-        public Contract IsLowerThan(int val, int comparer, string property, string message)
+        public Contract IsLowerThan(long val, long comparer, string property, string message)
         {
             if (val >= comparer)
                 AddNotification(property, message);
@@ -129,7 +129,7 @@
         #endregion
 
         #region IsLowerOrEqualsThan
-        public Contract IsLowerOrEqualsThan(decimal val, int comparer, string property, string message)
+        public Contract IsLowerOrEqualsThan(decimal val, long comparer, string property, string message)
         {
             if ((double)val > comparer)
                 AddNotification(property, message);
@@ -137,7 +137,7 @@
             return this;
         }
 
-        public Contract IsLowerOrEqualsThan(double val, int comparer, string property, string message)
+        public Contract IsLowerOrEqualsThan(double val, long comparer, string property, string message)
         {
             if (val > comparer)
                 AddNotification(property, message);
@@ -145,7 +145,7 @@
             return this;
         }
 
-        public Contract IsLowerOrEqualsThan(float val, int comparer, string property, string message)
+        public Contract IsLowerOrEqualsThan(float val, long comparer, string property, string message)
         {
             if (val > comparer)
                 AddNotification(property, message);
@@ -153,7 +153,7 @@
             return this;
         }
 
-        public Contract IsLowerOrEqualsThan(long val, int comparer, string property, string message)
+        public Contract IsLowerOrEqualsThan(int val, long comparer, string property, string message)
         {
             if (val > comparer)
                 AddNotification(property, message);
@@ -161,7 +161,7 @@
             return this;
         }
 
-        public Contract IsLowerOrEqualsThan(int val, int comparer, string property, string message)
+        public Contract IsLowerOrEqualsThan(long val, long comparer, string property, string message)
         {
             if (val > comparer)
                 AddNotification(property, message);
@@ -171,7 +171,7 @@
         #endregion
 
         #region AreEquals
-        public Contract AreEquals(decimal val, int comparer, string property, string message)
+        public Contract AreEquals(decimal val, long comparer, string property, string message)
         {
             if ((double)val != comparer)
                 AddNotification(property, message);
@@ -179,7 +179,7 @@
             return this;
         }
 
-        public Contract AreEquals(double val, int comparer, string property, string message)
+        public Contract AreEquals(double val, long comparer, string property, string message)
         {
             if (val != comparer)
                 AddNotification(property, message);
@@ -187,7 +187,7 @@
             return this;
         }
 
-        public Contract AreEquals(float val, int comparer, string property, string message)
+        public Contract AreEquals(float val, long comparer, string property, string message)
         {
             if (val != comparer)
                 AddNotification(property, message);
@@ -195,7 +195,7 @@
             return this;
         }
 
-        public Contract AreEquals(long val, int comparer, string property, string message)
+        public Contract AreEquals(int val, long comparer, string property, string message)
         {
             if (val != comparer)
                 AddNotification(property, message);
@@ -203,7 +203,7 @@
             return this;
         }
 
-        public Contract AreEquals(int val, int comparer, string property, string message)
+        public Contract AreEquals(long val, long comparer, string property, string message)
         {
             if (val != comparer)
                 AddNotification(property, message);
@@ -213,7 +213,7 @@
         #endregion
 
         #region AreNotEquals
-        public Contract AreNotEquals(decimal val, int comparer, string property, string message)
+        public Contract AreNotEquals(decimal val, long comparer, string property, string message)
         {
             if ((double)val == comparer)
                 AddNotification(property, message);
@@ -221,7 +221,7 @@
             return this;
         }
 
-        public Contract AreNotEquals(double val, int comparer, string property, string message)
+        public Contract AreNotEquals(double val, long comparer, string property, string message)
         {
             if (val == comparer)
                 AddNotification(property, message);
@@ -229,7 +229,7 @@
             return this;
         }
 
-        public Contract AreNotEquals(float val, int comparer, string property, string message)
+        public Contract AreNotEquals(float val, long comparer, string property, string message)
         {
             if (val == comparer)
                 AddNotification(property, message);
@@ -237,7 +237,7 @@
             return this;
         }
 
-        public Contract AreNotEquals(long val, int comparer, string property, string message)
+        public Contract AreNotEquals(int val, long comparer, string property, string message)
         {
             if (val == comparer)
                 AddNotification(property, message);
@@ -245,7 +245,7 @@
             return this;
         }
 
-        public Contract AreNotEquals(int val, int comparer, string property, string message)
+        public Contract AreNotEquals(long val, long comparer, string property, string message)
         {
             if (val == comparer)
                 AddNotification(property, message);
@@ -255,7 +255,7 @@
         #endregion
 
         #region Between      
-        public Contract IsBetween(int val, int from, int to, string property, string message)
+        public Contract IsBetween(long val, long from, long to, string property, string message)
         {
             if (!(val > from && val < to))
                 AddNotification(property, message);
